refactor: move osm auth init to separate file, use Depends
